requirements.txt
toga>=0.4.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python
numba

# GPU acceleration (optional — install the one matching your hardware)
# NVIDIA CUDA:   pip install cupy-cuda12x   (or cupy-cuda11x / cupy-cuda117)
# Any OpenCL:    pip install pyopencl
